+++ src/main.c
@@ -1,20 +1,20 @@
-#include "../pregame/main.h"
-#include "../debug.h"
-#include "../init/argv.h"
-#include "../init/bank.h"
-#include "../init/death.h"
-#include "../init/debug.h"
-#include "../init/graphics.h"
-#include "../init/kickout.h"
-#include "../init/monsters.h"
-#include "../init/stores.h"
-#include "../init/trade.h"
-#include "../io.h"
-#include "../kickout.h"
-#include "../master.h"
-#include "../player.h"
-#include "../variables.h"
-#include "main_loop.h"
+#include "pregame/main.h"
+#include "debug.h"
+#include "init/argv.h"
+#include "init/bank.h"
+#include "init/death.h"
+#include "init/debug.h"
+#include "init/graphics.h"
+#include "init/kickout.h"
+#include "init/monsters.h"
+#include "init/stores.h"
+#include "init/trade.h"
+#include "io.h"
+#include "kickout.h"
+#include "main_loop/main_loop.h"
+#include "master.h"
+#include "player.h"
+#include "variables.h"
 
 int main(const int argc, char *argv[]) {
