+++ src/main_loop/command.h
@@ -1,0 +1,3 @@
+#pragma once
+
+void command(long *command);
